Update to es6 syntax Breakout index.js to sub files in lib
